requirements.txt
streamlit>=1.32.0
supabase>=2.4.0
python-dotenv>=1.0.0
requests>=2.31.0
rapidfuzz>=3.0.0
recipe-scrapers>=14.55.0
safeway-api>=1.0.0
